+++ waficash/waficash/Models/User.cs
@@ -8,7 +8,7 @@ namespace waficash.Models
     {
         public long Id { get; set; }
         public string Name { get; set; }
-        public long AccountBalance { get; set; }
+        public decimal AccountBalance { get; set; }
         public long AccountNumber { get; set; }
         public string Currency { get; set; }
     }
